ejercicios/autogen/requirements.txt
pyautogen==0.2.29
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3
pillow==11.1.0
numpy==2.0.0
scikit-image==0.24.1
matplotlib==3.9.2
seaborn==0.13.2
jupyter==1.1.0
